# Back End/contracts/connected.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Safina connected app",
    "version": "0.6.0",
    "description": "Same-origin, cookie-authenticated API. Sandbox and live databases must be separate. Legacy v0.3 bearer API remains isolated."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "safina_session"
      }
    }
  },
  "paths": {
...
    "/program-rules": {
      "get": {
        "operationId": "get_program_rules",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/OperationalResult"
                }
              }
            }
          },
          "default": {
            "description": "Stable error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Error"
                }
              }
            }
          }
        }
      }
    },
    "/today": {
      "get": {
        "operationId": "get_today",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Today"
                }
              }
            }
          },
          "default": {
            "description": "Stable error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/calendar": {
      "get": {
        "operationId": "get_calendar",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Calendar"
                }
              }
            }
          },
          "default": {
            "description": "Stable error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/ship-progress": {
      "get": {
        "operationId": "get_ship_progress",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/ShipProgress"
                }
              }
            }
          },
          "default": {
            "description": "Stable error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Error"
                }
              }
            }
          }
        }
      }
    },
    "/khatmas": {
      "get": {
        "operationId": "get_khatmas",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Khatmas"
                }
              }
            }
          },
          "default": {
            "description": "Stable error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/quran/reference": {
      "get": {
        "operationId": "get_quran_reference",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/QuranReferenceVersion"
                }
              }
            }
          },
          "default": {
            "description": "Stable error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./connected.schema.json#/$defs/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}